Implement Texera User Avatars Featuer (#1373)
This PR creates an Angular component to show the avatar of a Texera user.

The library/component **ngx-avatar** does not support retrieving Google users' avatars by Google ID (https://github.com/HaithemMosbahi/ngx-avatar/issues/51), so I implemented the logic to retrieve Google users' avatars myself in the component. 

In order to send requests to Google People API, a public key is needed to provide. 
Creating a Google API Public Key is easy, please follow this instruction.
https://developers.google.com/maps/documentation/maps-static/get-api-key

Then fill the public key in the **environment.default.ts**
`google: {
    clientID: "",
    publicKey: "",
  },`

Demo: 
If the user is a normal Texera user, the avatar will be the default avatar with a random background color.
![normal user](https://user-images.githubusercontent.com/52941906/142727712-691254e9-d4ae-48f9-b70b-8a46bde91e54.gif)


If the user is a Google user, the avatar will be its Google profile picture. 
![google user](https://user-images.githubusercontent.com/52941906/142727715-3fc6e839-e33c-4405-9813-389cb573a152.gif)

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>