Change Login component to use ng-zorro instead of bootstrap. Remove unused components (#1470)
Currently we are using a mix of ng-zorro, boostrap, and angular material in our UI components. We want to gradually migrate to ng-zorro for a unified look and less maintenance effort. 

This PR migrates the login component to ng-zorro from bootstrap.


![Kapture 2022-03-08 at 17 41 29](https://user-images.githubusercontent.com/12578068/157356351-f07b8330-de31-4ee3-9056-b62c2e797df9.gif)


This PR also deleted some unused components in user dashboard. 
- two empty components: running-job-section and resource-section
- user dictionary related components, including user-dictionary upload. The user dictionary feature is not enabled and not used. Moreover, the newly added preset feature can server similar purposes.

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>