Add source in console message (#1752)
## New `source` field for `PythonConsoleMessage`
This PR adds a field `source` for `PythonConsoleMessage`, along with an option in the configuration menu to toggle the field's display.

This field can show where a message is generated. For example print messages will have `<module name>:<function name>:<line no>`:

<img width="936" alt="Screenshot 2022-11-30 at 03 55 16" src="https://user-images.githubusercontent.com/17627829/204789926-0c2177b3-b26f-474f-bb85-fbaa7ed98273.png">

## Optimization
This PR also makes `PythonConsoleUpdateEvent` (now renamed to `ConsoleUpdateEvent`) support sending multiple messages in one update. It is useful when the web server has a large diff of messages to be sent. The messages are sent in the order that the controller used to receive them.