Attribute change propagation (#1663)
In the past, when users deleted an attribute or changed an attribute name through an operator, Texera wouldn't propagate this change to the succeeding operators. This caused an error when users tried to run the workflow because the succeeding operators would contain invalid attributes. Additionally, in the case of changing the attribute name, it was inconvenient since users would have to change the attribute name in every related operator if they changed it in one place.

This PR will solve the above problem. It propagates the changes in attributes, i.e., input schema, the set of attributes, to succeeding operators. For example, when a user renames an attribute through an operator, the attribute name will be updated in all succeeding operators.

Co-authored-by: Zuozhi Wang <zuozhiw@gmail.com>