[Compiler Refactor #4] Introduce new APIs in OperatorDescriptor to support new OpExecConfig interface (#1796)
This PR is a follow-up of #1791 , where a new `OpExecConfig` interface
is introduced for the new workflow compiler layer. This PR changes the
`OperatorDescriptor` interface to use the new `OpExecConfig`. Currently,
the old and new interfaces exist in parallel. After all migrations are
completed, the old interface will be removed.

New interfaces:
```
  def newOperatorExecutor(operatorSchemaInfo: OperatorSchemaInfo): NewOpExecConfig = {
    throw new UnsupportedOperationException(
      "operator " + operatorIdentifier + " is not migrated to new OpExec API"
    )
  }

  // a logical operator corresponds multiple physical operators (a small DAG)
  def operatorExecutorMultiLayer(operatorSchemaInfo: OperatorSchemaInfo): PhysicalPlan = {
    new PhysicalPlan(List(newOperatorExecutor(operatorSchemaInfo)), List())
```

Logical operators that correpond to a single physical operator will
implement `newOperatorExecutor` interface.
Logical operators that correpond to a **multiple** physical operator
(such as aggregate) will implement `operatorExecutorMultiLayer`
interface.