Fix error overriding issue on the webserver (#1803)
This PR fixes an issue where the actual error is overridden by the "Key
not found" error thrown from `LogicalPlan`. When the schema propagation
of one operator throws an exception, earlier we silently printed the
error and returned None. Now we return a list of errors that occurred
during schema propagation and handle this list case by case.